templates/requirements/local.txt
# Local development dependencies go here
-r test.txt

# Deployment
Fabric

# Documentation
Sphinx==1.2.2

# Admin
django-debug-toolbar==1.2.1
